Move headers to crypto directory
rdrand.h is not a crytp header and is part of oecore

Addressed Simon's comments. keys.h is now sgxkeys.h, cleaned up host/main.c in tests
