+++ common/kdf.c
@@ -5,10 +5,10 @@
 #include <openenclave/bits/safecrt.h>
 #include <openenclave/bits/safemath.h>
 #include <openenclave/bits/types.h>
-#include <openenclave/internal/hmac.h>
+#include <openenclave/internal/crypto/hmac.h>
+#include <openenclave/internal/crypto/sha.h>
 #include <openenclave/internal/kdf.h>
 #include <openenclave/internal/raise.h>
-#include <openenclave/internal/sha.h>
 #include <openenclave/internal/utils.h>
 
 #include "common.h"
