+++ common/sgx/revocation.c
@@ -5,15 +5,16 @@
 #include <openenclave/bits/safecrt.h>
 #include <openenclave/bits/safemath.h>
 #include <openenclave/internal/calls.h>
-#include <openenclave/internal/cert.h>
-#include <openenclave/internal/crl.h>
+#include <openenclave/internal/crypto/cert.h>
+#include <openenclave/internal/crypto/crl.h>
+#include <openenclave/internal/crypto/ec.h>
+#include <openenclave/internal/crypto/sha.h>
 #include <openenclave/internal/datetime.h>
-#include <openenclave/internal/ec.h>
 #include <openenclave/internal/hexdump.h>
 #include <openenclave/internal/raise.h>
 #include <openenclave/internal/report.h>
 #include <openenclave/internal/sgxcertextensions.h>
-#include <openenclave/internal/sha.h>
+#include <openenclave/internal/thread.h>
 #include <openenclave/internal/trace.h>
 #include <openenclave/internal/utils.h>
 #include "../common.h"
