+++ common/sgx/sgxcertextensions.c
@@ -1,8 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 #include <openenclave/bits/safecrt.h>
-#include <openenclave/internal/cert.h>
-#include <openenclave/internal/ec.h>
+#include <openenclave/internal/crypto/cert.h>
+#include <openenclave/internal/crypto/ec.h>
 #include <openenclave/internal/hexdump.h>
 #include <openenclave/internal/raise.h>
 #include <openenclave/internal/sgxcertextensions.h>
