+++ enclave/crypto/asn1.c
@@ -7,7 +7,8 @@
 #include <mbedtls/oid.h>
 
 #include <openenclave/bits/safecrt.h>
-#include <openenclave/internal/asn1.h>
+#include <openenclave/corelibc/string.h>
+#include <openenclave/internal/crypto/asn1.h>
 #include <openenclave/internal/print.h>
 #include <openenclave/internal/raise.h>
 #include <openenclave/internal/utils.h>
