+++ enclave/crypto/crl.c
@@ -4,7 +4,8 @@
 #include "crl.h"
 #include <mbedtls/platform.h>
 #include <openenclave/bits/safecrt.h>
-#include <openenclave/internal/crl.h>
+#include <openenclave/corelibc/string.h>
+#include <openenclave/internal/crypto/crl.h>
 #include <openenclave/internal/print.h>
 #include <openenclave/internal/raise.h>
 #include <openenclave/internal/utils.h>
