+++ enclave/crypto/sha.c
@@ -4,9 +4,9 @@
 #include <mbedtls/sha256.h>
 
 #include <openenclave/bits/types.h>
+#include <openenclave/internal/crypto/sha.h>
 #include <openenclave/internal/defs.h>
 #include <openenclave/internal/raise.h>
-#include <openenclave/internal/sha.h>
 
 typedef struct _oe_sha256_context_impl
 {
