+++ enclave/tls_cert.c
@@ -5,11 +5,11 @@
 #include <openenclave/bits/safecrt.h>
 #include <openenclave/enclave.h>
 #include <openenclave/internal/cert.h>
+#include <openenclave/internal/crypto/sha.h>
 #include <openenclave/internal/print.h>
 #include <openenclave/internal/raise.h>
 #include <openenclave/internal/report.h>
 #include <openenclave/internal/sgxtypes.h>
-#include <openenclave/internal/sha.h>
 #include <openenclave/internal/utils.h>
 #include <stdio.h>
