+++ host/crypto/openssl/sha.c
@@ -2,9 +2,9 @@
 // Licensed under the MIT License.
 
 #include <openenclave/host.h>
+#include <openenclave/internal/crypto/sha.h>
 #include <openenclave/internal/defs.h>
 #include <openenclave/internal/raise.h>
-#include <openenclave/internal/sha.h>
 #include <openssl/sha.h>
 #include <stdio.h>
 #include <string.h>
