+++ include/openenclave/internal/crypto/hash.h
