+++ include/openenclave/internal/random.h
@@ -1,11 +1,10 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-#ifndef _OE_RANDOM_H
-#define _OE_RANDOM_H
+#ifndef _OE_RANDOM_INTERNAL_H
+#define _OE_RANDOM_INTERNAL_H
 
 #include <openenclave/bits/result.h>
-#include <openenclave/bits/types.h>
 
 OE_EXTERNC_BEGIN
 
@@ -25,4 +24,4 @@ oe_result_t oe_random_internal(void* data, size_t size);
 
 OE_EXTERNC_END
 
-#endif /* _OE_RANDOM_H */
+#endif /* _OE_RANDOM_INTERNAL_H */
