+++ tests/crypto/crl_tests.c
@@ -6,7 +6,7 @@
 #endif
 
 #include <openenclave/internal/cert.h>
-#include <openenclave/internal/crl.h>
+#include <openenclave/internal/crypto/crl.h>
 #include <openenclave/internal/hexdump.h>
 #include <openenclave/internal/tests.h>
 #include <stdarg.h>
