+++ tests/crypto/hmac_tests.c
@@ -5,7 +5,7 @@
 #include <openenclave/enclave.h>
 #endif
 
-#include <openenclave/internal/hmac.h>
+#include <openenclave/internal/crypto/hmac.h>
 #include <openenclave/internal/tests.h>
 #include <stdio.h>
 #include <string.h>
