+++ tests/crypto/sha_tests.c
@@ -5,7 +5,7 @@
 #include <openenclave/enclave.h>
 #endif
 
-#include <openenclave/internal/sha.h>
+#include <openenclave/internal/crypto/sha.h>
 #include <openenclave/internal/tests.h>
 #include <stdio.h>
 #include <string.h>
